123
123

+++ src/main/java/orders/Orders_Mgr.java
@@ -67,7 +67,7 @@ public class Orders_Mgr {
 	            bean.setOrder_no(this.rs.getInt("order_no"));
 	            bean.setOrder_time(this.rs.getTimestamp("order_time"));
 	            bean.setOrder_foods(this.rs.getString("order_foods"));
-	            bean.setOrder_price(this.rs.getLong("order_price"));
+	            bean.setOrder_price(this.rs.getInt("order_price"));
 	            bean.setOrder_discount(this.rs.getLong("order_discount"));
 	            bean.setOrder_coupon(this.rs.getString("order_coupon"));
 	            bean.setOrder_type(this.rs.getString("order_type"));
@@ -106,7 +106,7 @@ public class Orders_Mgr {
 	            bean.setOrder_no(this.rs.getInt("order_no"));
 	            bean.setOrder_time(this.rs.getTimestamp("order_time"));
 	            bean.setOrder_foods(this.rs.getString("order_foods"));
-	            bean.setOrder_price(this.rs.getLong("order_price"));
+	            bean.setOrder_price(this.rs.getInt("order_price"));
 	            bean.setOrder_discount(this.rs.getLong("order_discount"));
 	            bean.setOrder_coupon(this.rs.getString("order_coupon"));
 	            bean.setOrder_type(this.rs.getString("order_type"));
@@ -139,7 +139,7 @@ public class Orders_Mgr {
 	            bean.setOrder_no(this.rs.getInt("order_no"));
 	            bean.setOrder_time(this.rs.getTimestamp("order_time"));
 	            bean.setOrder_foods(this.rs.getString("order_foods"));
-	            bean.setOrder_price(this.rs.getLong("order_price"));
+	            bean.setOrder_price(this.rs.getInt("order_price"));
 	            bean.setOrder_discount(this.rs.getLong("order_discount"));
 	            bean.setOrder_coupon(this.rs.getString("order_coupon"));
 	            bean.setOrder_type(this.rs.getString("order_type"));
@@ -171,7 +171,7 @@ public class Orders_Mgr {
 	            bean.setOrder_no(this.rs.getInt("order_no"));
 	            bean.setOrder_time(this.rs.getTimestamp("order_time"));
 	            bean.setOrder_foods(this.rs.getString("order_foods"));
-	            bean.setOrder_price(this.rs.getLong("order_price"));
+	            bean.setOrder_price(this.rs.getInt("order_price"));
 	            bean.setOrder_discount(this.rs.getLong("order_discount"));
 	            bean.setOrder_coupon(this.rs.getString("order_coupon"));
 	            bean.setOrder_type(this.rs.getString("order_type"));
@@ -219,7 +219,7 @@ public class Orders_Mgr {
 	            bean.setOrder_no(this.rs.getInt("order_no"));
 	            bean.setOrder_time(this.rs.getTimestamp("order_time"));
 	            bean.setOrder_foods(this.rs.getString("order_foods"));
-	            bean.setOrder_price(this.rs.getLong("order_price"));
+	            bean.setOrder_price(this.rs.getInt("order_price"));
 	            bean.setOrder_discount(this.rs.getLong("order_discount"));
 	            bean.setOrder_coupon(this.rs.getString("order_coupon"));
 	            bean.setOrder_type(this.rs.getString("order_type"));
@@ -390,7 +390,7 @@ public class Orders_Mgr {
 				  	bean.setOrder_no(rs.getInt("order_no"));
 		            bean.setOrder_time(rs.getTimestamp("order_time"));
 		            bean.setOrder_foods(rs.getString("order_foods"));
-		            bean.setOrder_price(rs.getLong("order_price"));
+		            bean.setOrder_price(rs.getInt("order_price"));
 		            bean.setOrder_discount(rs.getLong("order_discount"));
 		            bean.setOrder_coupon(rs.getString("order_coupon"));
 		            bean.setOrder_type(rs.getString("order_type"));
